Ajoute un deuxième CTA pour envoyer des emails sur la page résultat (#2782)
* Ajoute un deuxième CTA pour envoyer des emails sur la page résultat

* Cache le bouton

* Affiche le bouton email en mobile et change le texte

* Change le wording du bouton
