+++ src/components/offline-results.vue
@@ -1,143 +1,33 @@
 <template>
-  <div v-if="!submitResult.ok" class="aj-offline-results">
+  <div v-if="show" class="aj-offline-results">
     <h2 class="aj-question"> Je garde ces informations&nbsp;! </h2>
     <span>
       Vous pouvez enregistrer les résultats de votre simulation pour les
       consulter plus tard.
     </span>
-
-    <Modal tag="span" analytics-category="Email">
-      <template #message>
-        <div class="aj-offline-results-button">
-          <button
-            type="button"
-            class="button primary text-center"
-            @click="reset"
-          >
-            <i class="fa fa-envelope-o" aria-hidden="true" />
-            Recevoir par email
-          </button>
-        </div>
-      </template>
-      <h2 class="aj-question"> Recevoir un récapitulatif par email </h2>
-      <p>
-        Si vous le souhaitez nous pouvons vous recontacter à deux reprises pour
-        faire le point sur les démarches que vous avez faites et les blocages
-        que vous avez rencontrés.
-      </p>
-      <div v-if="submitResult?.ok">
-        <i class="fa fa-check" />
-        On vous envoie un email&nbsp;!
-      </div>
-
-      <div v-if="submitResult?.error">
-        <i class="fa fa-exclamation-triangle" />
-        Une erreur s'est produite
-      </div>
-
-      <div v-if="submitResult?.waiting">
-        <i class="fa fa-spinner fa-spin" aria-hidden="true" />
-      </div>
-
-      <form
-        v-if="
-          submitResult &&
-          !(submitResult.ok || submitResult.waiting || submitResult.error)
-        "
-        ref="form"
-      >
-        <label for="email" class="form__group">Votre email</label>
-        <input id="email" v-model="email" type="email" name="email" required />
-        <WarningMessage v-if="errorMessage"
-          >Une adresse email valide doit être indiquée.</WarningMessage
-        >
-        <div class="aj-feedback-buttons">
-          <button
-            type="submit"
-            class="button outline text-center"
-            @click.prevent="getRecap(true)"
-          >
-            J'accepte d'être recontacté·e par email
-          </button>
-          <button
-            type="submit"
-            class="button outline red text-center"
-            @click.prevent="getRecap(false)"
-          >
-            Non merci, je préfère ne recevoir que le récapitulatif
-          </button>
-        </div>
-      </form>
-    </Modal>
+    <div class="aj-offline-results-button">
+      <SendRecapEmailButton class="primary"></SendRecapEmailButton>
+    </div>
+    <RecapEmailModal :id="id"></RecapEmailModal>
   </div>
 </template>
 
 <script>
-import axios from "axios"
-
-import Modal from "@/components/modal"
-import WarningMessage from "@/components/warning-message"
+import RecapEmailModal from "@/components/modals/recap-email-modal"
+import SendRecapEmailButton from "@/components/buttons/send-recap-email-button"
 
 export default {
   name: "OfflineResults",
   components: {
-    WarningMessage,
-    Modal,
+    RecapEmailModal,
+    SendRecapEmailButton,
   },
   props: {
     id: String,
   },
-  data: function () {
-    return {
-      email: undefined,
-      submitResult: {
-        ok: undefined,
-        waiting: undefined,
-        error: undefined,
-      },
-      errorMessage: undefined,
-    }
-  },
-  methods: {
-    reset: function () {
-      this.submitResult = {
-        ok: undefined,
-        waiting: undefined,
-        error: undefined,
-      }
-    },
-    getRecap: function (surveyOptin) {
-      if (!this.$refs.form.checkValidity()) {
-        this.errorMessage = true
-        this.$matomo &&
-          this.$matomo.trackEvent(
-            "General",
-            "Invalid form",
-            this.$route.fullPath
-          )
-        return
-      }
-
-      const uri = `api/simulation/${this.id}/followup`
-      const payload = {
-        email: this.email,
-        surveyOptin,
-      }
-
-      this.submitResult.ok = false
-      this.submitResult.error = false
-      this.submitResult.waiting = true
-      return axios
-        .post(uri, payload)
-        .then(() => {
-          this.submitResult.ok = true
-        })
-        .catch(() => {
-          this.submitResult.error = true
-        })
-        .finally(() => {
-          this.submitResult.waiting = false
-        })
+  computed: {
+    show() {
+      return this.$store.state.recapEmailState !== "ok"
     },
   },
 }
