+++ src/styles/partials/mobile/_ui-kit.scss
@@ -3,7 +3,7 @@ html {
     .aj-main-container {
       padding: 0;
       .aj-category-title-wrapper {
-        padding: 16px 0;
+        padding: 16px;
       }
 
       .aj-box-wrapper {
