+++ src/styles/partials/tablet/_ui-kit.scss
@@ -9,7 +9,12 @@ html {
     }
     .aj-main-container {
       .aj-category-title-wrapper {
+        flex-direction: column;
         padding: 16px 32px;
+
+        .recap-email-button {
+          margin-top: 16px;
+        }
       }
       .aj-box-wrapper {
         margin-bottom: 78px;
